requirements.txt
python-dotenv
sqlalchemy
pyodbc
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
langchain-ollama
huggingface-hub
sentence-transformers
torch
chromadb
scikit-learn
fastapi
uvicorn
jinja2
requests
pydantic
typing-extensions
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
